requirements.txt
# Core dependencies
anthropic==0.22.0
streamlit==1.36.0
PyMuPDF==1.23.8
python-docx==1.1.0
markdown==3.5.2

# Data processing
pandas==2.2.0
numpy==1.26.3

# Machine learning and NLP
transformers==4.37.2
torch==2.1.2
sentencepiece==0.1.99

# Visualization
matplotlib==3.8.2
altair==5.2.0
networkx==3.2.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
typing-extensions==4.9.0
pyyaml==6.0.1
